fix: 코칭ID 1~5 규칙 및 거리 int화

+++ analysis-service/src/main/kotlin/com/parkit/analysis/coaching/dto/CoachingSocketDto.kt
@@ -4,10 +4,10 @@ package com.parkit.analysis.coaching.dto
  * socket service로 전달될 코칭 알림 이벤트
  */
 data class ObstacleDistancesDto(
-	val frontDistance: Double,
-	val backDistance: Double,
-	val leftDistance: Double,
-	val rightDistance: Double,
+	val frontDistance: Int,
+	val backDistance: Int,
+	val leftDistance: Int,
+	val rightDistance: Int,
 )
 
 data class CoachingSocketDto(
